+++ src/App.scss
@@ -6,6 +6,7 @@ html {
   font-family: Arial, "Helvetica Neue", Helvetica, sans-serif;
 }
 
+
 .TodoInfo {
   width: max-content;
   padding: 8px;
