+++ src/components/UserInfo/UserInfo.jsx
@@ -1,4 +1,12 @@
 // Add the required props
-export const UserInfo = () => (
-  <>UserInfo markup</>
-);
+export const UserInfo = ({ user }) => {
+  const { name, email } = user;
+
+  return (
+    <>
+      <a className="UserInfo" href={`mailto:${email}`}>
+        {name}
+      </a>
+    </>
+  );
+};
