+++ src/components/function.js
@@ -1,0 +1,13 @@
+
+import todosFromServer from '../api/todos.json';
+import usersFromServer from '../api/users.json';
+
+function getUserById(userId) {
+  return usersFromServer.find(user => user.id === userId)
+    || null;
+}
+
+export const todos = todosFromServer.map(todo => ({
+  ...todo,
+  user: getUserById(todo.userId),
+}));
